+++ sccfm_cli/commands/inventory/devices/asa/user/change_password/__init__.py
@@ -1,0 +1,5 @@
+__all__ = ["AsaChangePasswordCommand"]
+
+from sccfm_cli.commands.inventory.devices.asa.user.change_password.command import (
+    AsaChangePasswordCommand,
+)