+++ sccfm_cli/commands/objects/network/command.py
@@ -6,8 +6,10 @@ import click
 from rich.console import Console
 
 from sccfm_cli.commands.base import BaseCommand
-from sccfm_cli.commands.objects.network.create import NetworkCreateCommand
-from sccfm_cli.commands.objects.network.delete import NetworkDeleteCommand
+from sccfm_cli.commands.objects.network.create import CreateNetworkObjectCommand
+from sccfm_cli.commands.objects.network.delete import DeleteNetworkObjectCommand
+from sccfm_cli.commands.objects.network.list import ListNetworkObjectCommand
+from sccfm_cli.commands.objects.network.update import UpdateNetworkObjectCommand
 
 
 class NetworkCommand(BaseCommand):
@@ -16,8 +18,10 @@ class NetworkCommand(BaseCommand):
     def __init__(self, console: Console) -> None:
         super().__init__(console)
         self._subcommands: List[BaseCommand] = [
-            NetworkCreateCommand(console),
-            NetworkDeleteCommand(console),
+            CreateNetworkObjectCommand(console),
+            ListNetworkObjectCommand(console),
+            UpdateNetworkObjectCommand(console),
+            DeleteNetworkObjectCommand(console),
         ]
 
     @property