+++ sccfm_cli/commands/objects/network/list/__init__.py
@@ -1,0 +1,3 @@
+from sccfm_cli.commands.objects.network.list.command import ListNetworkObjectCommand
+
+__all__ = ["ListNetworkObjectCommand"]