+++ sccfm_cli/commands/objects/network/update/__init__.py
@@ -1,0 +1,3 @@
+from sccfm_cli.commands.objects.network.update.command import UpdateNetworkObjectCommand
+
+__all__ = ["UpdateNetworkObjectCommand"]