+++ sccfm_core/services/__init__.py
@@ -2,15 +2,21 @@ from sccfm_core.services.health_service import HealthService, HealthStatus
 from sccfm_core.services.inventory import (
     AsaCommandLineService,
     AsaDiskFileService,
+    AsaUserPasswordService,
     InventoryService,
 )
-from sccfm_core.services.object_management import NetworkObjectService
+from sccfm_core.services.object_management import (
+    NetworkObjectListResponse,
+    NetworkObjectService,
+)
 
 __all__ = [
     "AsaCommandLineService",
     "AsaDiskFileService",
+    "AsaUserPasswordService",
     "HealthService",
     "HealthStatus",
     "InventoryService",
+    "NetworkObjectListResponse",
     "NetworkObjectService",
 ]